Debug Heroku

+++ SQL/SQL_management.py
@@ -20,7 +20,8 @@ def setup_db(app):
 
 
 def db_drop_and_create():
-    db.drop_all()
+    if 'postgres://' not in pre_DATABASE_URL:
+        db.drop_all()
     db.create_all()
